Seperate concerns of tests
