+++ tests/DataObjectGetStaticReturnTypeExtensionTest.php
@@ -2,16 +2,6 @@
 
 namespace SilbinaryWolf\SilverstripePHPStan\Tests;
 
-use PHPStan\PhpDoc;
-use PHPStan\Analyser\Scope;
-use PHPStan\Cache\Cache;
-use PHPStan\File\FileHelper;
-use PHPStan\PhpDoc\PhpDocStringResolver;
-use PHPStan\Type\FileTypeMapper;
-use PHPStan\Analyser\NodeScopeResolver;
-use PHPStan\Analyser\TypeSpecifier;
-use PHPStan\Reflection\BrokerAwareExtension;
-
 // SilverStripe
 use DataObject;
 use DataList;
@@ -23,11 +13,14 @@ use Injector;
 use MySQLDatabase;
 use MySQLPDODatabase;
 
-final class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
-{
-    /** @var \PHPStan\Broker\Broker */
-    private $broker;
+// NOTE(Jake): 2018-04-05
+//
+// Workaround Composer Autoloader / PHPUnit not working as I expect.
+//
+include_once(dirname(__FILE__).'/ResolverTest.php');
 
+class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
+{
     public function dataDynamicMethodReturnTypeExtensions(): array
     {
         return [
@@ -61,21 +54,6 @@ final class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
                 sprintf(SiteTree::class),
                 sprintf('%s::get_one(%s::class)', DataObject::class, SiteTree::class),
             ],
-            // Injector::inst()->get(File::class)
-            [
-                sprintf(File::class),
-                sprintf('%s::inst()->get(%s::class)', Injector::class, File::class),
-            ],
-            // Test `Injector::inst()->get('Cookie_Backend)` returns `CookieJar` (uses direct value in YML, SS 3.6.X)
-            [
-                sprintf('%s', CookieJar::class),
-                sprintf('%s::inst()->get(%s::class)', Injector::class, Cookie_Backend::class),
-            ],
-            // Test `Injector::inst()->get("MySQLPDODatabase")` returns `MySQLDatabase` (uses "class" array in YML, SS 3.6.X)
-            [
-                sprintf('%s', MySQLDatabase::class),
-                sprintf('%s::inst()->get(%s::class)', Injector::class, MySQLPDODatabase::class),
-            ],
             // 
             /*[
                 sprintf('%s', Object::class),
@@ -94,9 +72,7 @@ final class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
         string $expression
     )
     {
-        $dynamicMethodReturnTypeExtensions = [
-            new \SilbinaryWolf\SilverstripePHPStan\InjectorReturnTypeExtension(),
-        ];
+        $dynamicMethodReturnTypeExtensions = [];
         $dynamicStaticMethodReturnTypeExtensions = [
             new \SilbinaryWolf\SilverstripePHPStan\DataObjectGetStaticReturnTypeExtension(),
         ];
@@ -106,50 +82,6 @@ final class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
             $expression,
             $dynamicMethodReturnTypeExtensions, 
             $dynamicStaticMethodReturnTypeExtensions
-        );
-    }
-
-    public function dataFunctionReturnTypeExtensions(): array
-    {
-        return [
-            // Test `singleton('File)` returns `File`
-            [
-                sprintf('%s', File::class),
-                sprintf('singleton("%s")', File::class),
-            ],
-            // Test `singleton("Cookie_Backend")` returns `CookieJar` (uses direct value in YML, SS 3.6.X)
-            [
-                sprintf('%s', CookieJar::class),
-                sprintf('singleton("%s")', Cookie_Backend::class),
-            ],
-            // Test `singleton("MySQLPDODatabase")` returns `MySQLDatabase` (uses "class" array in YML, SS 3.6.X)
-            [
-                sprintf('%s', MySQLDatabase::class),
-                sprintf('singleton("%s")', MySQLPDODatabase::class),
-            ]
-        ];
-    }
-
-    /**
-     * @dataProvider dataFunctionReturnTypeExtensions
-     * @param string $description
-     * @param string $expression
-     */
-    public function testFunctionReturnTypeExtensions(
-        string $description,
-        string $expression
-    )
-    {
-        $dynamicFunctionReturnTypeExtensions = [
-            new \SilbinaryWolf\SilverstripePHPStan\SingletonReturnTypeExtension(),
-        ];
-        $this->assertTypes(
-            __DIR__ . '/data/data-object-dynamic-method-return-types.php',
-            $description,
-            $expression,
-            [],
-            [],
-            $dynamicFunctionReturnTypeExtensions
         );
     }
 }
